refactor: define MessagingPlatform in core
