+++ src/core/inbound-message.ts
@@ -1,4 +1,4 @@
-import type { MessagingPlatform } from '../platforms/types.js';
+import type { MessagingPlatform } from './messaging-platform.js';
 
 /**
  * Normalized inbound message.
