+++ src/core/messaging-platform.ts
@@ -1,0 +1,1 @@
+export type MessagingPlatform = 'whatsapp' | 'discord' | 'slack' | 'teams';
